# `perf`: Import CLI command modules asynchronously