--- a/.changeset/light-lies-guess.md
+++ b/.changeset/light-lies-guess.md
@@ -1,0 +1,5 @@
+---
+'sku': patch
+---
+
+CLI: Speed up command execution by asynchronously loading commands at runtime