--- a/packages/sku/src/program/commands/build-ssr/build-ssr.command.ts
+++ b/packages/sku/src/program/commands/build-ssr/build-ssr.command.ts
@@ -1,12 +1,12 @@
 import { Command } from 'commander';
 import { statsOption } from '../../options/stats/stats.option.js';
-import { buildSsrAction } from './build-ssr.action.js';
 
-export const buildSsrCommand = new Command('build-ssr');
-
-buildSsrCommand
+export const buildSsrCommand = new Command('build-ssr')
   .description(
     'Create a production build of a server-side rendered application.',
   )
   .addOption(statsOption)
-  .action(buildSsrAction);
+  .action(async (options) => {
+    const { buildSsrAction } = await import('./build-ssr.action.js');
+    await buildSsrAction(options);
+  });