--- a/packages/sku/src/program/commands/build/build.command.ts
+++ b/packages/sku/src/program/commands/build/build.command.ts
@@ -1,16 +1,14 @@
 import { Command } from 'commander';
 import { statsOption } from '../../options/stats/stats.option.js';
-import { buildAction } from './build.action.js';
 import { convertLoadableOption } from '@/program/options/convertLoadable/convertLoadable.option.js';
 
-const buildCommand = new Command('build');
-
-buildCommand
+export const buildCommand = new Command('build')
   .description(
     'Create a production build of a statically-rendered application.',
   )
   .addOption(statsOption)
   .addOption(convertLoadableOption)
-  .action(buildAction);
-
-export { buildCommand };
+  .action(async (options) => {
+    const { buildAction } = await import('./build.action.js');
+    await buildAction(options);
+  });