--- a/packages/sku/src/program/commands/configure/configure.command.ts
+++ b/packages/sku/src/program/commands/configure/configure.command.ts
@@ -1,8 +1,10 @@
 import { Command } from 'commander';
-import { configureAction } from '@/program/commands/configure/configure.action.js';
 
 export const configureCommand = new Command('configure');
 
 configureCommand
   .description('Emit and update configuration files for your project.')
-  .action(configureAction);
+  .action(async (options) => {
+    const { configureAction } = await import('./configure.action.js');
+    configureAction(options);
+  });