--- a/packages/sku/src/program/commands/format/format.command.ts
+++ b/packages/sku/src/program/commands/format/format.command.ts
@@ -1,9 +1,11 @@
 import { Command } from 'commander';
-import { formatAction } from './format.action.js';
 
 export const formatCommand = new Command('format');
 
 formatCommand
   .description('Apply all available lint and formatting fixes.')
   .argument('[paths...]', 'paths to format')
-  .action(formatAction);
+  .action(async (paths, options) => {
+    const { formatAction } = await import('./format.action.js');
+    await formatAction(paths, options);
+  });