--- a/packages/sku/src/program/commands/init/init.command.ts
+++ b/packages/sku/src/program/commands/init/init.command.ts
@@ -1,7 +1,6 @@
 import { Command } from 'commander';
 import { packageManagerOption } from '../../options/packageManager/packageManager.option.js';
 import { setPackageManager } from '../../../services/packageManager/context/packageManager.js';
-import { initAction } from '@/program/commands/init/init.action.js';
 
 export const initCommand = new Command('init');
 
@@ -16,4 +15,7 @@ initCommand
   .on('option:package-manager', (packageManager) => {
     setPackageManager(packageManager);
   })
-  .action(initAction);
+  .action(async (projectName, options) => {
+    const { initAction } = await import('./init.action.js');
+    await initAction(projectName, options);
+  });