--- a/packages/sku/src/program/commands/lint/lint.command.ts
+++ b/packages/sku/src/program/commands/lint/lint.command.ts
@@ -1,9 +1,11 @@
 import { Command } from 'commander';
-import { lintAction } from './lint.action.js';
 
 export const lintCommand = new Command('lint');
 
 lintCommand
   .description('Run lint tooling over your code.')
   .argument('[paths...]', 'paths to lint')
-  .action(lintAction);
+  .action(async (paths, options) => {
+    const { lintAction } = await import('./lint.action.js');
+    await lintAction(paths, options);
+  });