--- a/packages/sku/src/program/commands/pre-commit/pre-commit.command.ts
+++ b/packages/sku/src/program/commands/pre-commit/pre-commit.command.ts
@@ -1,8 +1,10 @@
 import { Command } from 'commander';
-import { preCommitAction } from '@/program/commands/pre-commit/pre-commit.action.js';
 
 export const preCommitCommand = new Command('pre-commit');
 
 preCommitCommand
   .description('Run the sku pre-commit hook.')
-  .action(preCommitAction);
+  .action(async (options) => {
+    const { preCommitAction } = await import('./pre-commit.action.js');
+    await preCommitAction(options);
+  });