--- a/packages/sku/src/program/commands/serve/serve.command.ts
+++ b/packages/sku/src/program/commands/serve/serve.command.ts
@@ -4,7 +4,6 @@ import {
   portOption,
   strictPortOption,
 } from '../../options/port/port.option.js';
-import { serveAction } from './serve.action.js';
 
 export const serveCommand = new Command('serve');
 
@@ -16,6 +15,7 @@ serveCommand
   .addOption(portOption)
   .addOption(strictPortOption)
   .action(async ({ site, skuContext }, command) => {
+    const { serveAction } = await import('./serve.action.js');
     const { environment } = command.optsWithGlobals();
     serveAction({
       site,