--- a/packages/sku/src/program/commands/start-ssr/start-ssr.action.ts
+++ b/packages/sku/src/program/commands/start-ssr/start-ssr.action.ts
@@ -1,7 +1,5 @@
 import type { StatsChoices } from '../../options/stats/stats.option.js';
 import type { SkuContext } from '@/context/createSkuContext.js';
-import { viteStartSsrHandler } from '@/program/commands/start-ssr/vite-start-ssr-handler.js';
-import { webpackStartSsrHandler } from '@/program/commands/start-ssr/webpack-start-ssr-handler.js';
 
 export const startSsrAction = async ({
   stats,
@@ -11,8 +9,12 @@ export const startSsrAction = async ({
   skuContext: SkuContext;
 }) => {
   if (skuContext.bundler === 'vite') {
+    const { viteStartSsrHandler } = await import('./vite-start-ssr-handler.js');
     viteStartSsrHandler(skuContext);
   } else {
+    const { webpackStartSsrHandler } = await import(
+      './webpack-start-ssr-handler.js'
+    );
     webpackStartSsrHandler({ stats, skuContext });
   }
 };