--- a/packages/sku/src/program/commands/start-ssr/start-ssr.command.ts
+++ b/packages/sku/src/program/commands/start-ssr/start-ssr.command.ts
@@ -1,5 +1,4 @@
 import { Command } from 'commander';
-import { startSsrAction } from './start-ssr.action.js';
 import { statsOption } from '@/program/options/stats/stats.option.js';
 
 const startSsrCommand = new Command('start-ssr');
@@ -9,6 +8,9 @@ startSsrCommand
     'Start the sku development server for a server-rendered application.',
   )
   .addOption(statsOption)
-  .action(startSsrAction);
+  .action(async (options) => {
+    const { startSsrAction } = await import('./start-ssr.action.js');
+    await startSsrAction(options);
+  });
 
 export { startSsrCommand };