--- a/packages/sku/src/program/commands/start/start.action.ts
+++ b/packages/sku/src/program/commands/start/start.action.ts
@@ -1,8 +1,6 @@
 import type { StatsChoices } from '@/program/options/stats/stats.option.js';
 import type { SkuContext } from '@/context/createSkuContext.js';
-import { viteStartHandler } from '@/program/commands/start/vite-start-handler.js';
 import type { Command } from 'commander';
-import { webpackStartHandler } from '@/program/commands/start/webpack-start-handler.js';
 
 export const startAction = async (
   {
@@ -16,8 +14,10 @@ export const startAction = async (
 ) => {
   const { environment } = command.optsWithGlobals();
   if (skuContext.bundler === 'vite') {
+    const { viteStartHandler } = await import('./vite-start-handler.js');
     viteStartHandler(skuContext);
   } else {
+    const { webpackStartHandler } = await import('./webpack-start-handler.js');
     webpackStartHandler({ stats, environment, skuContext });
   }
 };