--- a/packages/sku/src/program/commands/translations/commands/compile/compile.command.ts
+++ b/packages/sku/src/program/commands/translations/commands/compile/compile.command.ts
@@ -1,12 +1,14 @@
 import { Command } from 'commander';
 import { watchOption } from '../../../../options/watch/watch.option.js';
-import { compileAction } from './compile.action.js';
 
 const compileCommand = new Command('compile');
 
 compileCommand
   .description('Compile translations defined in .vocab directories.')
   .addOption(watchOption)
-  .action(compileAction);
+  .action(async (options) => {
+    const { compileAction } = await import('./compile.action.js');
+    await compileAction(options);
+  });
 
 export { compileCommand };