--- a/packages/sku/src/program/commands/translations/translations.command.ts
+++ b/packages/sku/src/program/commands/translations/translations.command.ts
@@ -1,12 +1,10 @@
 import { Command } from 'commander';
 import { commands } from './commands/index.js';
 
-const translationsCommand = new Command('translations');
-
-translationsCommand.description('Manage vocab translations');
+export const translationsCommand = new Command('translations').description(
+  'Manage vocab translations',
+);
 
 for (const command of commands) {
   translationsCommand.addCommand(command);
 }
-
-export { translationsCommand };