--- a/packages/sku/src/program/index.ts
+++ b/packages/sku/src/program/index.ts
@@ -1,6 +1,3 @@
-import fs from 'node:fs';
-import { dirname, resolve } from 'node:path';
-import { fileURLToPath } from 'node:url';
 import { Command } from 'commander';
 import { commands } from './commands/index.js';
 import { debugOption } from './options/debug/debug.option.js';
@@ -9,16 +6,11 @@ import { environmentOption } from './options/environment/environment.option.js';
 import { initDebug } from '@/utils/debug.js';
 import { experimentalBundlerOption } from '@/program/options/expirementalBundler/experimentalBundler.option.js';
 import { preActionHook } from '@/program/hooks/preAction/preAction.hook.js';
+import packageJson from 'sku/package.json' with { type: 'json' };
 
-const __dirname = dirname(fileURLToPath(import.meta.url));
+const { name, description, version } = packageJson;
 
-export const program = new Command();
-
-const { name, description, version } = JSON.parse(
-  fs.readFileSync(resolve(__dirname, '../../package.json'), 'utf-8'),
-);
-
-program
+export const program = new Command()
   .name(name)
   .description(description)
   .version(version)