Since we moved to `commander.js` our CLI has gotten a bit slower, with the tradeoff being a more flexible and powerful API. However, it turns out there was still a bit of perf on the table.

By asynchronously importing only the action module we need at runtime, we can shave off ~1.5s from sku's startup time. This is quite similar to how the sku CLI used to work when it was just duct tape and `require`s.

This performance improvement should be easy to verify visually by just running commands locally in some fixtures, but I did a few benchmarks too:

```md
`hyperfine 'na sku build'` inside the `public-path` fixture
  - `master`: 7s +- 0.5s
  - `async-cli-actions`: 5.5s +- 0.3s

`hyperfine 'na sku --version` inside `sku-test` fixture
  - `master`: 4.1s +- 0.05s
  - `async-cli-actions`: 2.25s +- 0.04s
```

I also did a similar thing for the build and start actions, so for example `vite start` and `webpack start` only import the modules they need. This doesn't really make much of a perf difference, but IMO it's still worth doing just to prevent any kind of side-effect pollution from occurring.

**EDIT:** This _might_ speed up CI test times since we use the CLI in a lot of tests, but our test time is quite variable so it's hard to tell.